renamed logger.py -> artifact_manager.py
